+++ packages/qb/pqb/src/columns/index.ts
@@ -6,6 +6,7 @@ export * from './number';
 export * from './string';
 export * from './dateTime';
 export * from './boolean';
+export * from './domain';
 export * from './enum';
 export * from './json';
 export * from './array';