+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # orchid-core
 
+## 0.1.4
+
+### Patch Changes
+
+- Support domain types
+
 ## 0.1.3
 
 ### Patch Changes